styling and date formatting for user bookings

+++ frontend/src/functions/convertDates.js
@@ -1,3 +1,5 @@
+import formatDateString from "./formatDateString"
+
 // This function takes in a date and converts it to a string
 // that is more readable for the user
 
@@ -8,7 +10,7 @@
 const convertDates = (date) => {
   const start = new Date(date)
   start.setDate(start.getDate() + 1)
-  const startMonth = start.toLocaleString('default', { month: 'long' })
+  const startMonth = start.toLocaleString('default', { month: 'short' })
   const startDay = start.getDate()
   const startYear = start.getFullYear()
   const startString = `${startMonth} ${startDay}, ${startYear}`
